#18: DataList: Selenium Wrapper (basic support)
