+++ primefaces-selenium-components/src/main/java/org/primefaces/extensions/selenium/component/base/AbstractPageableData.java
@@ -2,6 +2,9 @@ package org.primefaces.extensions.selenium.component.base;
 
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
+import org.primefaces.extensions.selenium.PrimeExpectedConditions;
+import org.primefaces.extensions.selenium.PrimeSelenium;
+import org.primefaces.extensions.selenium.component.model.data.Page;
 import org.primefaces.extensions.selenium.component.model.data.Paginator;
 
 import java.util.List;
@@ -17,5 +20,20 @@ public abstract class AbstractPageableData extends AbstractComponent {
         return paginator;
     }
 
-    public abstract Paginator getPaginator();
+    public Paginator getPaginator() {
+        return new Paginator(getPaginatorWebElement());
+    }
+
+    public void selectPage(Page page) {
+        page.getWebElement().click();
+        PrimeSelenium.waitGui().until(PrimeExpectedConditions.jQueryNotActive());
+    }
+
+    public void selectPage(int number) {
+        for (Page page : getPaginator().getPages()) {
+            if (page.getNumber() == number) {
+                selectPage(page);
+            }
+        }
+    }
 }
